Fix performance issues during system update process There were two major performance issues that are now fixed in this release.
The first issue was to do with getting the updates and performing APT
package index searches to find all package dependencies and to calculate
the estimated update size: `system_executeCommand` was being performed
in parallel, which massively slowed down the system when upwards of 30
instances were running at the same time. This was fixed by using a
`Promise` chain, and also by taking advantage of being able to specify
multiple package names in `apt-cache show`.

The second issue was to do with the privileged interface, whereby the
update status and progress communicator was constantly sending messages
to the Settings app, slowing the Settings app down due to becoming
swamped with messages. This was fixed by batching status updates so that
they are only sent every 500 ms.

This means that the auto restart countdown timer shows up correctly,
instead of the Settings app update status lagging behind by a good 10
seconds, and therefore the system restarting with very little notice.